Bottom navigation added with folder structure
