+++ app/build.gradle.kts
@@ -54,7 +54,7 @@ dependencies {
     //Firebase
     implementation(platform("com.google.firebase:firebase-bom:32.2.3"))
     implementation("com.google.firebase:firebase-analytics-ktx")
-
+    implementation("com.google.firebase:firebase-database-ktx")
     //Maps
     implementation ("com.google.android.gms:play-services-maps:18.1.0")
