+++ app/src/main/java/com/example/capstone/model/User.kt
@@ -1,0 +1,6 @@
+package com.example.capstone.model
+
+data class User(
+    val email : String? = null,
+    val userName : String? = null
+)
