+++ app/src/main/java/com/example/capstone/ui/ProfileFragment.kt
@@ -1,0 +1,20 @@
+package com.example.capstone.ui
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.capstone.R
+
+class ProfileFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+
+
+        return inflater.inflate(R.layout.fragment_profile, container, false)
+    }
+}
