+++ app/src/main/java/com/example/capstone/utils/Constant.kt
@@ -1,4 +1,4 @@
-package com.example.capstone
+package com.example.capstone.utils
 
 class Constant {
     companion object {
