+++ app/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#ffffff"
-    tools:context=".HomeActivity">
+    tools:context=".ui.HomeActivity">
 
     <FrameLayout
         android:id="@+id/container"
