+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,9 @@
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
 
+
+    <string name="profile">Profile</string>
+    <string name="home">Home</string>
+    <string name="logout">Logout</string>
+
 </resources>
